requirements.txt
# Core Dependencies
aiogram==3.3.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
requests==2.31.0
asyncio==3.4.3
aiohttp==3.9.1
jinja2==3.1.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
asyncpg==0.29.0

# AI/ML Dependencies
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
prophet==1.1.5
ta==0.10.2
yfinance==0.2.22

# OpenAI & NLP
openai==1.3.7
transformers==4.36.2
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# Crypto & Trading
# ccxt==4.1.77
# python-binance==1.0.19

# Security & Encryption
# cryptography==41.0.8
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Web scraping & Data
beautifulsoup4==4.12.2
tweepy==4.14.0
praw==7.7.1

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
APScheduler==3.10.4
pydantic==2.5.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk==1.38.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
